% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hybrid_phe.R
\docType{data}
\name{hybrid_phe}
\alias{hybrid_phe}
\title{Phenotypic data of hybrids}
\format{A data frame with 410 rows and 3 variables:
\describe{
  \item{\code{M}}{The names of male parents.}
  \item{\code{F}}{The names of female parents.}
  \item{\code{GY}}{The grain yield of hybrids.}
}}
\usage{
hybrid_phe
}
\description{
This dataset contains phenotypic data of 410 hybrids for grain yield in maize.
}
\keyword{datasets}
