% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_geno.R
\docType{data}
\name{input_geno}
\alias{input_geno}
\title{Genotype in Hapmap Format}
\format{A data frame with 4979 rows and 359 columns.}
\usage{
input_geno
}
\description{
Genotypic data of 348 maize inbred lines in Hapmap format with double bit.
}
\keyword{datasets}
