% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/input_geno1.R
\docType{data}
\name{input_geno1}
\alias{input_geno1}
\title{Genotype in Numeric Format}
\format{A data frame with 1000 rows and 50 variables.}
\usage{
input_geno1
}
\description{
Genotypic data of 50 rice inbred lines with 1000 SNPs.
}
\keyword{datasets}
